fix: remove dhclient "hack" from rc.local on DPUs
## Description
This was put in place as part of https://redmine.mellanox.com/issues/3531747,
which we confirmed was fixed 2 years ago as part of HBN 1.5 (which is documented
in the Redmine itself).

In the time that has passed, we've been running two `dhclient` instances,
which didn't really present itself as an issue until recently, when we started
rolling out DOCA 2.9.3.

See https://nvidia.slack.com/archives/C02RKLCN8BT/p1759209121722859 for
more information (this was me troubleshooting + investigating + looking into
what was going on, including the +1 to get rid of this hack).

In DOCA 2.9.3, DHCP lease management has switched from `dhclient` over to
`systemd-networkd`, so we now have two completely separate processes managing
the `oob_net0` lease.

Generally this is "fine", since both will get the same IP back from `carbide-dhcp`,
and the `oob_net0` interface will have the correct singular IP, but we end up
with two completely different systems managing updating the lease, AND, in the case
of `qa3`, we ran into an issue where there was a mismatch in lease information,
and the `oob_net0` interface had two IPs on it. I think that might be a whole
separate issue wrt the incorrect IP, but step 1 is to get rid of this hack.

There will be a separate MR to also add `duppet` config to drop this file
from already allocated instances as well.

## Type of Change
<!-- Check one that best describes this MR -->
- [ ] **Add** - New feature or capability
- [ ] **Change** - Changes in existing functionality  
- [x] **Fix** - Bug fixes
- [ ] **Remove** - Removed features or deprecated functionality
- [x] **Internal** - Internal changes (refactoring, tests, docs, etc.)

## Related Tickets (Optional)
NVBugs: [nbvugs-5548285](https://nvbugspro.nvidia.com/bug/5548285)

## Breaking Changes
- [ ] This MR contains breaking changes

<!-- If checked above, describe the breaking changes and migration steps -->

## Testing
<!-- How was this tested? Check all that apply -->
- [ ] Unit tests added/updated
- [ ] Integration tests added/updated  
- [x] Manual testing performed
- [ ] No testing required (docs, internal refactor, etc.)

## Additional Notes
https://nvidia.slack.com/archives/C02RKLCN8BT/p1759209121722859

See merge request nvmetal/carbide!4662