feat: add FileEnsure support to duppet so we can delete files too
## Description
*The reason behind this change is https://jirasw.nvidia.com/browse/FORGE-7062,
because we need to clean up an old `/etc/rc.local` hack we had in place that
is causing both `dhclient` and `networkd` to both manage DHCP leases as of the
latest DOCA 2.9.3 release. Another MR is going in alongside this to remove the
hack from the image: !4662,
but we also want to proactively clean up the `/etc/rc.local` file from existing
allocations.

The idea was to leverage `duppet`, which totally makes sense, and goes back to
a previous conversation a few of us had that we should add the ability to delete
files, so here we are!
*

This adds a new `FileEnsure` capability to `duppet`, with the options being:
- `FileEnsure::Present`
- `FileEnsure::Absent`

The default is `FileEnsure::Present`.

This also changes the constructor interface a little bit to chain together
some builder-pattern methods to build the spec, including:
- `.with_content(..)`
- `.with_perms(..)`
- `.with_ownership(..)`
- `.with_ensure(..)`

Tested by:
- Writing unit tests.
- `scp`'ing a build of this to a DPU and running `sudo forge-dpu-agent duppet --dry-run` (followed by a run without `--dry-run`) to make sure it cleaned up `/etc/rc.local`.

## Type of Change
<!-- Check one that best describes this MR -->
- [x] **Add** - New feature or capability
- [ ] **Change** - Changes in existing functionality  
- [ ] **Fix** - Bug fixes
- [ ] **Remove** - Removed features or deprecated functionality
- [x] **Internal** - Internal changes (refactoring, tests, docs, etc.)

## Related Tickets (Optional)
NVBugs: [nvbugs-5548285](https://nvbugspro.nvidia.com/bug/5548285)

## Breaking Changes
- [ ] This MR contains breaking changes

<!-- If checked above, describe the breaking changes and migration steps -->

## Testing
<!-- How was this tested? Check all that apply -->
- [x] Unit tests added/updated
- [ ] Integration tests added/updated  
- [x] Manual testing performed
- [ ] No testing required (docs, internal refactor, etc.)

## Additional Notes
Long live duppet!

See merge request nvmetal/carbide!4665