Alter install script to only gracefully reboot bf2 bmc
## Description

The bf3 bmc is more resilient than the bf2 bmc and will upgrade itself if the host loses power. This means we can rely on the next steps in the process that power off the host on bf3 and save time in the install. This commit removes the additional graceful reboot of the dpu bmc during the installation for bf3s.

## Type of Change

<!--Check one that best describes this MR-->

- [ ] **Add** - New feature or capability
- [ ] **Change** - Changes in existing functionality
- [ ] **Fix** - Bug fixes
- [ ] **Remove** - Removed features or deprecated functionality
- [x] **Internal** - Internal changes (refactoring, tests, docs, etc.)

## Related Tickets (Optional)

<!--If applicable, provide JIRA or/and NVBugs ticket. If none, the MR itself will be referenced in changelog-->

JIRA: NVBugs:

## Breaking Changes

- [ ] This MR contains breaking changes

<!--If checked above, describe the breaking changes and migration steps-->

## Testing

<!--How was this tested? Check all that apply-->

- [ ] Unit tests added/updated
- [ ] Integration tests added/updated
- [ ] Manual testing performed
- [ ] No testing required (docs, internal refactor, etc.)

## Additional Notes

<!--Any additional context, deployment notes, or reviewer guidance-->

See merge request nvmetal/carbide!4663