carbide-api split: Move all models out of db and into model
## Description

Make the `model` module the home for all "model-like" things, including database objects, in preparation for `model` and `db` being separate crates. The `db` module now contains only private types and "bare" functions that operate on models.

This is the first step in the process of splitting the carbide-api crate into multiple smaller crates. The rough plan is:

- Move models out of DB (you are here)
- Drop all usages of CarbideError out of the model module (involves some moving of CarbideError variants into RpcDataConversionError)
- Convert the model module to a crate
- Convert the db module to a crate
- Convert the state controllers and site explorer to crate(s)
- Convert the GRPC handlers and web stuff to crates

## Type of Change
<!-- Check one that best describes this MR -->
- [ ] **Add** - New feature or capability
- [ ] **Change** - Changes in existing functionality  
- [ ] **Fix** - Bug fixes
- [ ] **Remove** - Removed features or deprecated functionality
- [X] **Internal** - Internal changes (refactoring, tests, docs, etc.)

## Breaking Changes
- [ ] This MR contains breaking changes

## Testing
<!-- How was this tested? Check all that apply -->
- [ ] Unit tests added/updated
- [ ] Integration tests added/updated  
- [ ] Manual testing performed
- [X] No testing required (docs, internal refactor, etc.)

## Additional Notes

One downside of this is that the `model` module contains `FromRow` implementations as well as some of the `PgJson` types. This is an unfortunate consequence of rust's orphaning rules: Any FromRow/TryFrom/etc on these types need to live in the same crate as the types themselves, so in preparation for the split-out, they have to live in `model`. An alternative is using NewTypes everywhere in `db` but I'm not sure that's worth the extra boilerplate right now.

See merge request nvmetal/carbide!4673