Log errors on auth_callback
## Description
Currently if there are any errors in `/admin/auth-callback`, we return error messages in the response body, but on our end we just see `HTTP status: 500 Internal Server Error` in the logs. Fix that by using a custom error type and logging any errors that happen. 

## Type of Change
<!-- Check one that best describes this MR -->
- [ ] **Add** - New feature or capability
- [ ] **Change** - Changes in existing functionality  
- [X] **Fix** - Bug fixes
- [ ] **Remove** - Removed features or deprecated functionality
- [ ] **Internal** - Internal changes (refactoring, tests, docs, etc.)

## Breaking Changes
- [ ] This MR contains breaking changes

## Testing
<!-- How was this tested? Check all that apply -->
- [ ] Unit tests added/updated
- [ ] Integration tests added/updated  
- [ ] Manual testing performed
- [X] No testing required (docs, internal refactor, etc.)

## Additional Notes
Not tested since we don't have a great story around mocking the oauth flow. This is one of those "if it compiles, it works" situations IMO.

See merge request nvmetal/carbide!4685