Enable plugins and exec in skaffold config
## Description
CI builds are failing due to a missing ksops plugin.

The carbide base module in forged now has a secretGenerator that requires ksops. This requires skaffold to pass the right options to kustomize to enable plugins and exec.

## Type of Change
- [ ] **Add** - New feature or capability
- [ ] **Change** - Changes in existing functionality  
- [ ] **Fix** - Bug fixes
- [ ] **Remove** - Removed features or deprecated functionality
- [X] **Internal** - Internal changes (refactoring, tests, docs, etc.)

## Related Tickets (Optional) 

## Breaking Changes
- [ ] This MR contains breaking changes

## Testing
- [ ] Unit tests added/updated
- [ ] Integration tests added/updated  
- [ ] Manual testing performed
- [X] No testing required (docs, internal refactor, etc.)

## Additional Notes
If this doesn't work, I can bring back the code which generates the secret as part of the `setup-k3s-env-ips` step in forged (which I had originally done to avoid needing to enable plugins in skaffold)

See merge request nvmetal/carbide!4693