carbide-api split: Move NextState code out of models
## Description

As part of splitting carbide_api::model into its own crate, the various state controller states in model/controller_outcome.rs need to stay as models, since they're persisted to the database and communicated over RPC.

But the business logic of the state controller code itself should stay in the same crate as the rest of the state controllers: The immediate reason is that it uses StateHandlerError, which needs to stay with state_controller, but also it's business logic specific to state handling, and not specific to the modeling of states.

To do this we need to define traits like ManagedHostStateHelper, ReprovisionStateHelper, etc, which let us define methods like all_dpus_in_sync, next_bfb_install_state, etc inside state_handler.rs even when the types are moved to their own crate.


## Type of Change
<!-- Check one that best describes this MR -->
- [ ] **Add** - New feature or capability
- [ ] **Change** - Changes in existing functionality  
- [ ] **Fix** - Bug fixes
- [ ] **Remove** - Removed features or deprecated functionality
- [X] **Internal** - Internal changes (refactoring, tests, docs, etc.)

## Breaking Changes
- [ ] This MR contains breaking changes

## Testing
<!-- How was this tested? Check all that apply -->
- [ ] Unit tests added/updated
- [ ] Integration tests added/updated  
- [ ] Manual testing performed
- [X] No testing required (docs, internal refactor, etc.)

## Additional Notes
This is one of several MR's going up as a prelude to splitting api/src/model into its own crate, in order to make the final MR smaller.

See merge request nvmetal/carbide!4687