carbide-api split: Move parts of CarbideConfig into model
## Description
As part of splitting carbide_api::model into its own crate, there's some cross-dependencies between carbide_api::cfg::file and carbide_api::model, such that some models depend on config objects. For each of these cases, move the relevant types into api/src/model.

This is not an attempt to go through every type defined in CarbideConfig and decide for every one of them which should be a model, but is a pragmatic change to move the minimum amount over to break the circular dependency. Mainly, right now that means all the Firmware types, HostHealthConfig, and DpuModel.

## Type of Change
<!-- Check one that best describes this MR -->
- [ ] **Add** - New feature or capability
- [ ] **Change** - Changes in existing functionality  
- [ ] **Fix** - Bug fixes
- [ ] **Remove** - Removed features or deprecated functionality
- [X] **Internal** - Internal changes (refactoring, tests, docs, etc.)

## Breaking Changes
- [ ] This MR contains breaking changes

## Testing
<!-- How was this tested? Check all that apply -->
- [ ] Unit tests added/updated
- [ ] Integration tests added/updated  
- [ ] Manual testing performed
- [X] No testing required (docs, internal refactor, etc.)

## Additional Notes
This is one of several MR's going up as a prelude to splitting api/src/model into its own crate, in order to make the final MR smaller.

See merge request nvmetal/carbide!4689