Use cargo fmt's group_imports option for deterministic import ordering
## Description
Fixing merge conflicts with import statements is annoying, and so is reviewing tons of diff churn where import statements move around. This introduces using the `group_imports=StdExternaCrate` and `imports_granularity=module` options in cargo fmt, so that our imports are consistent all the time.

This should help reduce churn in diffs for `use` statements by telling cargo fmt to order imports with std::* first, then any external imports, then crate imports, and using a module-level import granularity.

It requires a nightly cargo though, which is annoying but can be solved via a Makefile.toml task and being baked into the Dockerfile. The nightly cargo will only be used for the check-format-nightly and format-nightly tasks. **The nightly cargo/rust is not used for anything else, just these two cargo-make tasks.**

## Type of Change
<!-- Check one that best describes this MR -->
- [ ] **Add** - New feature or capability
- [ ] **Change** - Changes in existing functionality  
- [ ] **Fix** - Bug fixes
- [ ] **Remove** - Removed features or deprecated functionality
- [X] **Internal** - Internal changes (refactoring, tests, docs, etc.)

## Breaking Changes
- [ ] This MR contains breaking changes

## Testing
- [ ] Unit tests added/updated
- [ ] Integration tests added/updated  
- [ ] Manual testing performed
- [X] No testing required (docs, internal refactor, etc.)

## Additional Notes
This is a draft to solicit feedback and see if the CI runner can cope with it.

See merge request nvmetal/carbide!4698