Added lvscan to the qcow-imager script.  Made grub update optional.
## Description

To allow tenants to use LVM devices for the root filesystem, this adds a lvscan before we try to find the device.

Additionally for the storage team, 2 new options are available on the linux command line.

update_grub_template **&** update_grub_cfg

They default to "yes" to maintain existing functionality, but can be set to "no" on the command line to disable.

## Type of Change

<!--Check one that best describes this MR-->

- [ ] **Add** - New feature or capability
- [ ] **Change** - Changes in existing functionality
- [x] **Fix** - Bug fixes
- [ ] **Remove** - Removed features or deprecated functionality
- [ ] **Internal** - Internal changes (refactoring, tests, docs, etc.)

## Related Tickets (Optional)

5533959 5563186

## Breaking Changes

- [ ] This MR contains breaking changes

<!--If checked above, describe the breaking changes and migration steps-->

## Testing

Tested using the instances stuck in provisioning state that the storage team were bringing up.

- [ ] Unit tests added/updated
- [ ] Integration tests added/updated
- [x] Manual testing performed
- [ ] No testing required (docs, internal refactor, etc.)

## Additional Notes

<!--Any additional context, deployment notes, or reviewer guidance-->

See merge request nvmetal/carbide!4641