carbide-api split: Split out model into its own crate
## Description
Now that the prelude MR's are merged, this actually splits api/src/model into a `carbide-api-model` crate, stored in api/model.  The crate's name is carbide-api-model, but it's just called "model" from code.

Most of this is simply rewriting imports/types for the new crate, except:

- Un-tag stuff as cfg(test), since we can't use cfg(test) for conditional compilations across crates

- Move some From<T> impl's to the model crate where needed

- Duplicate some of the include_bytes! consts where needed to avoid needing to import them

- Move the legacy TopologyDataV1/etc definitions to the relevant test modules to avoid needing to pollute the model crate with them

## Type of Change
- [ ] **Add** - New feature or capability
- [ ] **Change** - Changes in existing functionality  
- [ ] **Fix** - Bug fixes
- [ ] **Remove** - Removed features or deprecated functionality
- [X] **Internal** - Internal changes (refactoring, tests, docs, etc.)

## Breaking Changes
- [ ] This MR contains breaking changes

## Testing
- [ ] Unit tests added/updated
- [ ] Integration tests added/updated  
- [ ] Manual testing performed
- [X] No testing required (docs, internal refactor, etc.)

## Additional Notes
I have the db crate split out as well in a branch, see the `wip/carbide-api-split/06-db-crate-split` branch... I want this MR to merge first before I put up another half-dozen or so MR's as a prelude to that split-out.

See merge request nvmetal/carbide!4700