carbide-api split: Move resource_pool types into model crate and database features to db module
## Description
As part of making api::db its own crate, it needs to see types defined in the resource_pool module. Move the types to the model crate so db can see them, and move the db functionality into the db crate itself.

## Type of Change

- [ ] **Add** - New feature or capability
- [ ] **Change** - Changes in existing functionality  
- [ ] **Fix** - Bug fixes
- [ ] **Remove** - Removed features or deprecated functionality
- [X] **Internal** - Internal changes (refactoring, tests, docs, etc.)

## Breaking Changes
- [ ] This MR contains breaking changes

## Testing
- [ ] Unit tests added/updated
- [ ] Integration tests added/updated  
- [ ] Manual testing performed
- [X] No testing required (docs, internal refactor, etc.)

## Additional Notes
This is part of a series of MRs which are going up as a prelude to splitting api::db into its own crate.

See merge request nvmetal/carbide!4712