carbide-api split: Make api::db not dependent on api::Api
## Description
As part of making api::db its own crate, it needs to not depend on api::Api. There are a couple of places where we're passing an `Api` object to database functions instead of doing something finer-grained. Fix that by:

- calling db::resource_pool::allocate from within db::vpc::allocate_dpa_vni rather than calling back to Api::pool_allocate_dpa_vni
- Passing a transaction instead of an Api object if all we were using the Api object for was to get a transaction.

## Type of Change

- [ ] **Add** - New feature or capability
- [ ] **Change** - Changes in existing functionality  
- [ ] **Fix** - Bug fixes
- [ ] **Remove** - Removed features or deprecated functionality
- [X] **Internal** - Internal changes (refactoring, tests, docs, etc.)

## Breaking Changes
- [ ] This MR contains breaking changes

## Testing
- [ ] Unit tests added/updated
- [ ] Integration tests added/updated  
- [ ] Manual testing performed
- [X] No testing required (docs, internal refactor, etc.)

## Additional Notes
This is part of a series of MRs which are going up as a prelude to splitting api::db into its own crate.

See merge request nvmetal/carbide!4706