carbide-api split: Move CURRENT_STATE_MODEL_VERSION out of db crate
## Description
As part of making api::db its own crate, it can't depend back on state machine things like CURRENT_STATE_MODEL_VERSION. Instead, make it a parameter to the relevant functions.

## Type of Change

- [ ] **Add** - New feature or capability
- [ ] **Change** - Changes in existing functionality  
- [ ] **Fix** - Bug fixes
- [ ] **Remove** - Removed features or deprecated functionality
- [X] **Internal** - Internal changes (refactoring, tests, docs, etc.)

## Breaking Changes
- [ ] This MR contains breaking changes

## Testing
- [ ] Unit tests added/updated
- [ ] Integration tests added/updated  
- [ ] Manual testing performed
- [X] No testing required (docs, internal refactor, etc.)

## Additional Notes
This is part of a series of MRs which are going up as a prelude to splitting api::db into its own crate.

See merge request nvmetal/carbide!4708