Split some more small crates to avoid rebuilds
## Description
Currently, carbide-api depends on api-test-helper as a dev dependency,
because it uses it in integration tests. But api-test-helper depends on
carbide-api because its job is to make helpers to launch carbide-api.

The upshot of this is that any change to carbide-api, even a unit test,
triggers a rebuild of api-test-helper, which in turn rebuilds
carbide-api again.

Fix this by making carbide-api's integration tests their own crate
(again... this was the case a while back as well), and make
prometheus_text_parser its own crate, as it's a utility used both by
carbide-api's unit tests and ssh-console's integration tests.

## Type of Change
- [ ] **Add** - New feature or capability
- [ ] **Change** - Changes in existing functionality  
- [ ] **Fix** - Bug fixes
- [ ] **Remove** - Removed features or deprecated functionality
- [X] **Internal** - Internal changes (refactoring, tests, docs, etc.)

## Breaking Changes
- [ ] This MR contains breaking changes

## Testing
- [ ] Unit tests added/updated
- [ ] Integration tests added/updated  
- [ ] Manual testing performed
- [X] No testing required (docs, internal refactor, etc.)

## Additional Notes

See merge request nvmetal/carbide!4718