Add a script that copies the carbide repo while sanitizing the result
## Description
This is for sharing the carbide repo with mirantis, for which the
agreement is that we'll show them the carbide code but won't provide any
support. (The expectation is that they'll read the code but likely not
try to run it as-is)

This is a separate step from the eventual open-sourcing of carbide,
which will require other changes to the code first. The mirantis sharing
is much shorter term and merely requires a sanitized repo.

## Type of Change
- [ ] **Add** - New feature or capability
- [ ] **Change** - Changes in existing functionality  
- [ ] **Fix** - Bug fixes
- [ ] **Remove** - Removed features or deprecated functionality
- [X] **Internal** - Internal changes (refactoring, tests, docs, etc.)

## Related Tickets (Optional)

JIRA: [FORGE-7078](https://jirasw.nvidia.com/browse/FORGE-7078)

## Breaking Changes
- [ ] This MR contains breaking changes

## Testing
- [ ] Unit tests added/updated
- [ ] Integration tests added/updated  
- [ ] Manual testing performed
- [X] No testing required (docs, internal refactor, etc.)

## Additional Notes
The sanitized repo can be found at https://gitlab-master.nvidia.com/nvmetal/carbide-mirror-mirantis

I'm committing the script itself so that we can iterate on which things should be excluded, and anticipating that we may need to snapshot carbide again if there's changes we need to make in the code first.

See merge request nvmetal/carbide!4733