feat: Emit a metric on the amount of open logging/tracing spans
## Description

In order to get a better understanding of whether all spans are closed, this change adds a very simple tracing layer which increases a counter on span openening and closing.

The value of of the counter is emitted as metric `carbide_api_tracing_spans_open`.
Since the measurements are so simply, the feature is not behind a flag.  If a flag is required in the future, we could add it.

It would also be possible to count the amount of spans by span name, in case we determine there's a leak.

## Type of Change
- [x] **Add** - New feature or capability
- [ ] **Change** - Changes in existing functionality  
- [ ] **Fix** - Bug fixes
- [ ] **Remove** - Removed features or deprecated functionality
- [ ] **Internal** - Internal changes (refactoring, tests, docs, etc.)

## Related Tickets (Optional)

NVBugs: https://nvbugspro.nvidia.com/bug/5517703

## Breaking Changes
- [ ] This MR contains breaking changes

## Testing

- [x] Unit tests added/updated
- [ ] Integration tests added/updated  
- [ ] Manual testing performed
- [ ] No testing required (docs, internal refactor, etc.)

## Additional Notes

See merge request nvmetal/carbide!4748