move-deps-to-workspace: Cargo.toml fixup utility
## Description
This is a utility to make it so that all dependencies define their
versions in the workspace Cargo.toml. It uses toml_edit to take any
dependency defined in any Cargo.toml in the workspace, and moves its
version spec to the toplevel Cargo.toml, replacing the original spec
with `workspace = true`. If multiple crates specify different versions
of the same dependency, the newer version is used.

Defining dependency versions one time at the workspace root ensures we
don't needlessly depend on multiple versions of the same crate, which
helps to decrease bloat and increase consistency across the codebase.

This uses the ["xtask"][0] convention, which is supposed to be for
utilities used in a project like these.  The xtask crate uses the first
argument to determine what to do... I didn't bother using clap or
anything (yet) as there's just one supported task, and no cli flags yet.

Run with `cargo xtask move-deps-to-workspace`.

## Type of Change
- [ ] **Add** - New feature or capability
- [ ] **Change** - Changes in existing functionality  
- [ ] **Fix** - Bug fixes
- [ ] **Remove** - Removed features or deprecated functionality
- [X] **Internal** - Internal changes (refactoring, tests, docs, etc.)

## Breaking Changes
- [ ] This MR contains breaking changes

## Testing
- [ ] Unit tests added/updated
- [ ] Integration tests added/updated  
- [ ] Manual testing performed
- [X] No testing required (docs, internal refactor, etc.)

## Additional Notes

[0]: https://github.com/matklad/cargo-xtask

See merge request nvmetal/carbide!4767