carbide-oss: Drop storage management APIs
## Description
The whole storage system is currently not being worked on any more, and the
depenency on libnvmesh is an issue for open-sourcing carbide.

It's not being used in any site, and it's basically all dead code at this point.
So, delete:

- All RPC calls for storage-related things
- Database/model implementations for storage-related stuff
- Storage-related tables, including os_image_id column from instances table
- WaitingForStorageConfig state from machine state controller

## Type of Change
- [ ] **Add** - New feature or capability
- [ ] **Change** - Changes in existing functionality  
- [ ] **Fix** - Bug fixes
- [X] **Remove** - Removed features or deprecated functionality
- [ ] **Internal** - Internal changes (refactoring, tests, docs, etc.)

## Related Tickets (Optional)
JIRA: [FORGE-7309](https://jirasw.nvidia.com/browse/FORGE-7309) Remove nvmesh dependency

## Breaking Changes
- [X] This MR contains breaking changes

It breaks anyone trying to configure storage clusters, but nobody's doing that in any environment. So it shouldn't break anything.

## Testing
- [ ] Unit tests added/updated
- [ ] Integration tests added/updated  
- [X] Manual testing performed
- [ ] No testing required (docs, internal refactor, etc.)

## Additional Notes
Verified that I could deploy to a cluster already running current trunk, and that the migration was successful.

See merge request nvmetal/carbide!4782