Dont override explicit_start_needed if it is set to true
## Description

Currently, when multiple metadata.toml get merged into one, `explicit_start_needed` may ended up being False even if one of the metadata files have it set to True. This is problematic because it means that the firmware for that vendor and model will become automatic, rather than requiring explicit start.

The changes here make it so that it is set to True even if only one metadata.toml file has it set to True.

## Type of Change
<!-- Check one that best describes this MR -->
- [ ] **Add** - New feature or capability
- [x] **Change** - Changes in existing functionality  
- [ ] **Fix** - Bug fixes
- [ ] **Remove** - Removed features or deprecated functionality
- [ ] **Internal** - Internal changes (refactoring, tests, docs, etc.)


## Breaking Changes
- [ ] This MR contains breaking changes

<!-- If checked above, describe the breaking changes and migration steps -->

## Testing
<!-- How was this tested? Check all that apply -->
- [x] Unit tests added/updated
- [ ] Integration tests added/updated  
- [x] Manual testing performed
- [ ] No testing required (docs, internal refactor, etc.)

See merge request nvmetal/carbide!4806