Fix postgres bind limits for observations in testing
## Description
I configured a massive lo-ip pool to test ingesting thousands of hosts, and ran
into issues with too many query parameters.

For one, we need to make sure to divide the bind limit by the actual number of
items per tuple (5, not 4, in the case of resource_pool::populate)

For another, actually using 65535 query parameters still seems to break on my
machine, we need more headroom than that. Cut it in half just to be sure.

## Type of Change
- [ ] **Add** - New feature or capability
- [ ] **Change** - Changes in existing functionality  
- [X] **Fix** - Bug fixes
- [ ] **Remove** - Removed features or deprecated functionality
- [ ] **Internal** - Internal changes (refactoring, tests, docs, etc.)

## Breaking Changes
- [ ] This MR contains breaking changes

## Testing
<!-- How was this tested? Check all that apply -->
- [ ] Unit tests added/updated
- [ ] Integration tests added/updated  
- [X] Manual testing performed
- [ ] No testing required (docs, internal refactor, etc.)

## Additional Notes

See merge request nvmetal/carbide!4809