Make machine-a-tron put less load on carbide-api
## Description
During bringup of 5000 mock machines (plus 1 dpu each), the constant calls to
IdentifySerialRequest and DiscoverMachine cause carbide-api to slow down
immensely.

IdentifySerialRequest is only used to know in the TUI what the MachineId of a
mock machine is, for the brief interval between when site explorer created a
machine for it, but before it boots to scout or dpu-agent. If we skip it, it
makes no difference at all for non-TUI uses of machine-a-tron, and even for
people using the TUI, it only makes the machine take slightly longer to show the
correct ID in the TUI.

Then there's DiscoverMachine, which machine-a-tron does automatically when
simulating a machine booting, but this is too early if site-explorer hasn't
discovered the machine yet (which never the case in production.) If this
happens, just sleep forever until the next reboot rather than retrying over and
over.

With these changes, I can ingest 5,000 mock machines (10,000 endpoints) in less
than 20 minutes on my local machine.

## Type of Change
- [ ] **Add** - New feature or capability
- [ ] **Change** - Changes in existing functionality  
- [X] **Fix** - Bug fixes
- [ ] **Remove** - Removed features or deprecated functionality
- [ ] **Internal** - Internal changes (refactoring, tests, docs, etc.)

## Breaking Changes
- [ ] This MR contains breaking changes

## Testing
- [ ] Unit tests added/updated
- [ ] Integration tests added/updated  
- [X] Manual testing performed
- [ ] No testing required (docs, internal refactor, etc.)

## Additional Notes

See merge request nvmetal/carbide!4808