Faster IP allocation for machine interfaces
## Description

When running carbide with 10,000+ endpoints, the amount of time it takes to find
the next available IP starts increasing to several seconds by the time we're
over several thousand of them. Before long it becomes a major bottleneck.

Most of the complexity in IpAllocator comes from its ability to allocate whole
subnets (/30's, etc), which is currently never used (we always allocate /32's)
and will likely never be used for machine interfaces.

So, carve out a shortcut for allocating machine IP's where we get the next IP in
a subnet from a single postgres query. This makes allocation essentially
instantaneous.

## Type of Change
- [ ] **Add** - New feature or capability
- [ ] **Change** - Changes in existing functionality  
- [ ] **Fix** - Bug fixes
- [ ] **Remove** - Removed features or deprecated functionality
- [X] **Internal** - Internal changes (refactoring, tests, docs, etc.)

## Breaking Changes
- [ ] This MR contains breaking changes

## Testing
- [ ] Unit tests added/updated
- [X] Integration tests added/updated  
- [X] Manual testing performed
- [ ] No testing required (docs, internal refactor, etc.)

## Additional Notes

See merge request nvmetal/carbide!4802