chore: Move PowerOptions out of StateHandlerContext
## Description

This change moves the `power_options` parameter out of `StateHandlerContext`, with the goal of making the state controller framework agnostic to the workload it controls.

The behavior should be mostly as before. However a small difference is that the power options are now persisted before the main transaction, instead of after.

## Type of Change

- [ ] **Add** - New feature or capability
- [ ] **Change** - Changes in existing functionality  
- [ ] **Fix** - Bug fixes
- [ ] **Remove** - Removed features or deprecated functionality
- [x] **Internal** - Internal changes (refactoring, tests, docs, etc.)

## Related Tickets (Optional)

## Breaking Changes
- [ ] This MR contains breaking changes

## Testing

- [ ] Unit tests added/updated
- [ ] Integration tests added/updated  
- [ ] Manual testing performed
- [x] No testing required (docs, internal refactor, etc.)

## Additional Notes

See merge request nvmetal/carbide!4831